refactor(config): make cleaner config module
