+++ packages/shared/config/env-vars/index.js
@@ -1,5 +1,9 @@
-const node = require('./node');
+const frontAuth = require('./front-auth');
+const nodejs = require('./nodejs');
+const passport = require('./passport');
 
 module.exports = {
-  node,
+  frontAuth,
+  nodejs,
+  passport,
 };
