+++ packages/shared/config/index.js
@@ -1,7 +1,5 @@
 const envVars = require('./env-vars');
-const jwt = require('./jwt');
 
 module.exports = {
   envVars,
-  jwt,
 };
