--- packages/shared/config/jwt.js
@@ -1,6 +1,0 @@
-const { ExtractJwt } = require('passport-jwt');
-
-module.exports = {
-  jwtSecret: process.env.JWT_SECRET,
-  jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-};
